backend/pyproject.toml
[tool.poetry]
name = "strata"
version = "0.1.0"
description = "Strata. Your stuff, organized."
authors = ["fd"]
license = "GPL-3.0"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"

# --- Web framework ---
fastapi = ">=0.128.0,<0.129.0"
starlette = ">=0.50.0,<0.51.0"
uvicorn = ">=0.40.0,<0.41.0"

# --- Async stack ---
anyio = ">=4.12.1,<5.0.0"
h11 = ">=0.16.0,<0.17.0"
httptools = ">=0.7.1,<0.8.0"
uvloop = ">=0.22.1,<0.23.0"
watchfiles = ">=1.1.1,<2.0.0"
websockets = ">=16.0,<17.0"
idna = ">=3.11,<4.0"

# --- Data / validation ---
pydantic = ">=2.12.5,<3.0.0"
pydantic-core = ">=2.41.5,<3.0.0"
annotated-types = ">=0.7.0,<0.8.0"
typing-extensions = ">=4.15.0,<5.0.0"
typing-inspection = ">=0.4.2,<0.5.0"
annotated-doc = ">=0.0.4,<0.0.5"

# --- Database ---
sqlalchemy = ">=2.0.45,<3.0.0"
alembic = ">=1.18.0,<2.0.0"

# --- Configuration ---
python-dotenv = ">=1.2.1,<2.0.0"
python-decouple = ">=3.8,<4.0.0"
pyyaml = ">=6.0.3,<7.0.0"

# --- CLI ---
click = ">=8.3.1,<9.0.0"

[tool.poetry.scripts]
api = "app.main:app"
test = "app.cli:run"
test-coverage = "app.cli:coverage"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4"
pytest-cov = "^4.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
